Started riot window

+++ filebrowser/ImGuiFileBrowser.cpp
@@ -88,12 +88,12 @@ namespace imgui_addons
 
             //Render top file bar for easy navigation
             ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.882f, 0.745f, 0.078f,1.0f));
-            for(int i = 0; i < current_dirlist.size(); i++)
+            for(int i = 0; i < (int)current_dirlist.size(); i++)
             {
                 if( ImGui::Button(current_dirlist[i].c_str()) )
                 {
                     //If last button clicked, nothing happens
-                    if(i == current_dirlist.size() - 1)
+                    if(i == (int)current_dirlist.size() - 1)
                         break;
                     show_error |= !(onNavigationButtonClick(i));
                 }
@@ -101,7 +101,7 @@ namespace imgui_addons
                 //Draw Arrow Buttons
                 ImGui::PushStyleColor(ImGuiCol_Button, ImVec4(1.0f, 1.0f, 1.0f, 0.01f));
                 ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(1.0f, 1.0f, 1.0f,1.0f));
-                if(i != current_dirlist.size() - 1)
+                if(i != (int)current_dirlist.size() - 1)
                 {
                     ImGui::SameLine(0,0);
                     ImGui::ArrowButtonEx("##Right", ImGuiDir_Right, ImVec2(frame_height,frame_height), ImGuiButtonFlags_Disabled);
@@ -122,13 +122,13 @@ namespace imgui_addons
                 filter_dirty = false;
                 filtered_dirs.clear();
                 filtered_files.clear();
-                for(int i = 0; i < subdirs.size(); i++)
+                for(int i = 0; i < (int)subdirs.size(); i++)
                 {
                     if(filter.PassFilter(subdirs[i].name.c_str()))
                         filtered_dirs.push_back(&subdirs[i]);
                 }
 
-                for(int i = 0; i < subfiles.size(); i++)
+                for(int i = 0; i < (int)subfiles.size(); i++)
                 {
                     if(filter.PassFilter(subfiles[i].name.c_str()))
                         filtered_files.push_back(&subfiles[i]);
@@ -156,7 +156,7 @@ namespace imgui_addons
             #endif // OSWIN
 
             ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.882f, 0.745f, 0.078f,1.0f));
-            for (int i = 0; i < filtered_dirs.size(); i++)
+            for (int i = 0; i < (int)filtered_dirs.size(); i++)
             {
                 if(!filtered_dirs[i]->is_hidden || show_hidden)
                 {
@@ -175,7 +175,7 @@ namespace imgui_addons
             ImGui::PopStyleColor(1);
 
             //Output files
-            for (int i = 0; i < filtered_files.size(); i++)
+            for (int i = 0; i < (int)filtered_files.size(); i++)
             {
                 if(!filtered_files[i]->is_hidden || show_hidden)
                 {
@@ -274,12 +274,12 @@ namespace imgui_addons
 
             //Render top file bar for easy navigation
             ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.882f, 0.745f, 0.078f,1.0f));
-            for(int i = 0; i < current_dirlist.size(); i++)
+            for(int i = 0; i < (int)current_dirlist.size(); i++)
             {
                 if( ImGui::Button(current_dirlist[i].c_str()) )
                 {
                     //If last button clicked, nothing happens
-                    if(i == current_dirlist.size() - 1)
+                    if(i == (int)current_dirlist.size() - 1)
                         break;
                     show_error |= !(onNavigationButtonClick(i));
                 }
@@ -287,7 +287,7 @@ namespace imgui_addons
                 //Draw Arrow Buttons
                 ImGui::PushStyleColor(ImGuiCol_Button, ImVec4(1.0f, 1.0f, 1.0f, 0.01f));
                 ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(1.0f, 1.0f, 1.0f,1.0f));
-                if(i != current_dirlist.size() - 1)
+                if(i != (int)current_dirlist.size() - 1)
                 {
                     ImGui::SameLine(0,0);
                     ImGui::ArrowButtonEx("##Right", ImGuiDir_Right, ImVec2(frame_height,frame_height), ImGuiButtonFlags_Disabled);
@@ -317,7 +317,7 @@ namespace imgui_addons
 
             ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.882f, 0.745f, 0.078f,1.0f));
             int items = 0;
-            for (int i = 0; i < subdirs.size(); i++)
+            for (int i = 0; i < (int)subdirs.size(); i++)
             {
                 if(!subdirs[i].is_hidden || show_hidden)
                 {
@@ -336,7 +336,7 @@ namespace imgui_addons
             ImGui::PopStyleColor(1);
 
             //Output files
-            for (int i = 0; i < subfiles.size(); i++)
+            for (int i = 0; i < (int)subfiles.size(); i++)
             {
                 if(!subfiles[i].is_hidden || show_hidden)
                 {
@@ -377,7 +377,7 @@ namespace imgui_addons
             ImGui::PushItemWidth(80);
             if(ImGui::BeginCombo("##saveTypes", valid_exts[selected_ext_idx].c_str()))
             {
-                for(int i = 0; i < valid_exts.size(); i++)
+                for(int i = 0; i < (int)valid_exts.size(); i++)
                 {
                     if(ImGui::Selectable(valid_exts[i].c_str(), selected_ext_idx == i))
                     {
@@ -386,7 +386,7 @@ namespace imgui_addons
                         size_t idx = name.find_last_of(".");
                         if(idx == std::string::npos)
                             idx = strlen(save_fn);
-                        for(int i = 0; i < valid_exts[selected_ext_idx].size(); i++)
+                        for(int i = 0; i < (int)valid_exts[selected_ext_idx].size(); i++)
                             save_fn[idx++] = valid_exts[selected_ext_idx][i];
                         save_fn[idx++] = '\0';
                     }
@@ -638,7 +638,7 @@ namespace imgui_addons
             ImGui::SetCursorPosY(ImGui::GetCursorPosY() + 5);
             ImGui::SetNextWindowContentSize(ImVec2(0, 17 * valid_exts.size() ));
             ImGui::BeginChild("##SupportedExts", ImVec2(0, -35), true);
-            for(int i = 0; i < valid_exts.size(); i++)
+            for(int i = 0; i < (int)valid_exts.size(); i++)
                 ImGui::BulletText("%s", valid_exts[i].c_str());
             ImGui::EndChild();
